Scroll wheel moves sight Using float instead of double for expected/actuals.

+++ browser/src/commonMain/kotlin/client/Expect.kt
@@ -11,13 +11,13 @@ expect val windowAdapter: WindowAdapter
 expect val htmlCanvasElementAdapter: HTMLCanvasElementAdapter
 
 expect fun alert(message: String)
-expect suspend fun awaitAnimationFrame(): Double
-expect fun getDevicePixelRatio(): Double
+expect suspend fun awaitAnimationFrame(): Float
+expect fun getDevicePixelRatio(): Float
 expect fun getLocationHost(): String
 expect fun uint8ArrayAdapterFactory(length: Int): Uint8ArrayAdapter
 expect fun rtcPeerConnectionAdapterFactory(configuration: JsonObject): RTCPeerConnectionAdapter
 expect fun audioAdapterFactory(src: String): AudioAdapter
-expect fun setShellsStatusBar(percent: Double)
-expect fun setArmorStatusBar(percent: Double)
-expect fun setMinesStatusBar(percent: Double)
+expect fun setShellsStatusBar(percent: Float)
+expect fun setArmorStatusBar(percent: Float)
+expect fun setMinesStatusBar(percent: Float)
 expect fun setBuilderMode(builderMode: BuilderMode)
